make remove_readout_os an instance method of KData
