+++ src/mrpro/algorithms/__init__.py
@@ -1,2 +1,1 @@
 from mrpro.algorithms._prewhiten_kspace import prewhiten_kspace
-from mrpro.algorithms._remove_readout_os import remove_readout_os
