+++ src/mrpro/data/__init__.py
@@ -9,6 +9,7 @@ from mrpro.data._EncodingLimits import EncodingLimits
 from mrpro.data._EncodingLimits import Limits
 from mrpro.data._KHeader import KHeader
 from mrpro.data._Data import Data
+from mrpro.data._CsmData import CsmData
 from mrpro.data._kdata._KData import KData
 from mrpro.data._KNoise import KNoise
 from mrpro.data._IHeader import IHeader
@@ -16,4 +17,3 @@ from mrpro.data._IData import IData
 from mrpro.data._DcfData import DcfData
 from mrpro.data._QHeader import QHeader
 from mrpro.data._QData import QData
-from mrpro.data._CsmData import CsmData
