+++ src/mrpro/operators/_SensitivityOp.py
@@ -12,6 +12,8 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+from __future__ import annotations
+
 import torch
 
 from mrpro.data import CsmData
